+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to
 
 ### Changed
 
+- 🛂(backend) stop throttling collaboration servers #1730
 - 🚸(backend) use unaccented full name for user search #1637
 - 🌐(backend) internationalize demo #1644